[[ -f /usr/local/bin/apt ]] || { return ;  }

+++ bash_conf/009-inotify.sh
@@ -1,4 +1,5 @@
 [[ -f /usr/bin/inotifywait ]] || {
+   [[ -f /usr/local/bin/apt ]] || { return ;  }
    ct_install_inotify-tools() {
       sudo apt install inotify-tools
    }
